+++ seed/go-sdk/audiences/dynamic-snippets/example2/snippet.go
@@ -2,12 +2,17 @@ package example
 
 import (
     client "github.com/audiences/fern/client"
+    option "github.com/audiences/fern/option"
     context "context"
     fern "github.com/audiences/fern"
 )
 
 func do() () {
-    client := client.NewClient()
+    client := client.NewClient(
+        option.WithBaseURL(
+            "https://api.fern.com",
+        ),
+    )
     client.Foo.Find(
         context.TODO(),
         &fern.FindRequest{