+++ seed/go-sdk/cross-package-type-names/dynamic-snippets/example1/snippet.go
@@ -2,11 +2,16 @@ package example
 
 import (
     client "github.com/cross-package-type-names/fern/client"
+    option "github.com/cross-package-type-names/fern/option"
     context "context"
 )
 
 func do() () {
-    client := client.NewClient()
+    client := client.NewClient(
+        option.WithBaseURL(
+            "https://api.fern.com",
+        ),
+    )
     client.FolderA.Service.GetDirectThread(
         context.TODO(),
     )