+++ seed/go-sdk/examples/exported-client-name/dynamic-snippets/example6/snippet.go
@@ -9,6 +9,9 @@ import (
 
 func do() () {
     client := client.NewAcmeClient(
+        option.WithBaseURL(
+            "https://api.fern.com",
+        ),
         option.WithToken(
             "<token>",
         ),