+++ seed/go-sdk/examples/exported-client-name/dynamic-snippets/example9/snippet.go
@@ -8,6 +8,9 @@ import (
 
 func do() () {
     client := client.NewAcmeClient(
+        option.WithBaseURL(
+            "https://api.fern.com",
+        ),
         option.WithToken(
             "<token>",
         ),