+++ seed/go-sdk/extends/dynamic-snippets/example0/snippet.go
@@ -2,12 +2,17 @@ package example
 
 import (
     client "github.com/extends/fern/client"
+    option "github.com/extends/fern/option"
     context "context"
     fern "github.com/extends/fern"
 )
 
 func do() () {
-    client := client.NewClient()
+    client := client.NewClient(
+        option.WithBaseURL(
+            "https://api.fern.com",
+        ),
+    )
     client.ExtendedInlineRequestBody(
         context.TODO(),
         &fern.Inlined{