+++ seed/go-sdk/oauth-client-credentials/dynamic-snippets/example1/snippet.go
@@ -2,12 +2,17 @@ package example
 
 import (
     client "github.com/oauth-client-credentials/fern/client"
+    option "github.com/oauth-client-credentials/fern/option"
     context "context"
     fern "github.com/oauth-client-credentials/fern"
 )
 
 func do() () {
-    client := client.NewClient()
+    client := client.NewClient(
+        option.WithBaseURL(
+            "https://api.fern.com",
+        ),
+    )
     client.Auth.RefreshToken(
         context.TODO(),
         &fern.RefreshTokenRequest{