+++ seed/go-sdk/path-parameters/no-custom-config/dynamic-snippets/example3/snippet.go
@@ -2,11 +2,16 @@ package example
 
 import (
     client "github.com/path-parameters/fern/client"
+    option "github.com/path-parameters/fern/option"
     context "context"
 )
 
 func do() () {
-    client := client.NewClient()
+    client := client.NewClient(
+        option.WithBaseURL(
+            "https://api.fern.com",
+        ),
+    )
     client.User.GetUser(
         context.TODO(),
         "tenant_id",