+++ seed/go-sdk/response-property/dynamic-snippets/example5/snippet.go
@@ -2,11 +2,16 @@ package example
 
 import (
     client "github.com/response-property/fern/client"
+    option "github.com/response-property/fern/option"
     context "context"
 )
 
 func do() () {
-    client := client.NewClient()
+    client := client.NewClient(
+        option.WithBaseURL(
+            "https://api.fern.com",
+        ),
+    )
     client.Service.GetMovie(
         context.TODO(),
         "string",