+++ seed/ts-sdk/oauth-client-credentials-custom/src/api/resources/auth/client/requests/index.ts
@@ -1,0 +1,2 @@
+export { type GetTokenRequest } from "./GetTokenRequest";
+export { type RefreshTokenRequest } from "./RefreshTokenRequest";